requirements.txt
django>=3.0,<5.0
djangorestframework>=3.12,<4.0
djangorestframework-simplejwt>=4.8.0,<5.0.0
celery>=5.2,<6.0
redis>=4.0,<5.0
django-celery-beat>=2.0,<3.0
python-decouple>=3.5,<4.0
